feat: add useFetchBySlug, useFetchById hooks and plain js fetchBySlug and fetchById methods
