+++ packages/components/package.json
@@ -64,8 +64,8 @@
     "vite": "^4.4.9"
   },
   "peerDependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": ">=17.0.0",
+    "react-dom": ">=17.0.0"
   },
   "dependencies": {
     "@contentful/experience-builder-core": "file:../core",
