+++ packages/core/package.json
@@ -50,6 +50,7 @@
     "@typescript-eslint/eslint-plugin": "^6.12.0",
     "@typescript-eslint/parser": "^6.12.0",
     "@vitest/coverage-v8": "^1.0.4",
+    "contentful": "^10.6.4",
     "eslint": "^8.54.0",
     "happy-dom": "^12.10.3",
     "rimraf": "^5.0.5",
@@ -62,5 +63,8 @@
     "@contentful/rich-text-types": "^16.3.0",
     "@contentful/visual-sdk": "^1.0.0-alpha.34",
     "lodash-es": "^4.17.21"
+  },
+  "peerDependencies": {
+    "contentful": ">=10.6.0"
   }
 }
