+++ packages/core/src/entity/index.ts
@@ -1,1 +1,2 @@
 export * from './EditorModeEntityStore';
+export * from './EntityStore';
