+++ packages/core/src/fetchers/index.ts
@@ -1,0 +1,2 @@
+export * from './fetchBySlug';
+export * from './fetchById';
