+++ packages/core/src/index.ts
@@ -3,3 +3,4 @@ export * from './definitions/styles';
 export * from './entity';
 export * from './communication';
 export * from './enums';
+export * from './fetchers';
