+++ packages/experience-builder-sdk/package.json
@@ -31,7 +31,6 @@
     "@contentful/rich-text-types": "^16.2.1",
     "@contentful/visual-sdk": "^1.0.0-alpha.26",
     "classnames": "^2.3.2",
-    "contentful": "^10.6.10",
     "csstype": "^3.1.2",
     "lodash.get": "^4.4.2",
     "lodash.omit": "^4.5.0",
@@ -59,6 +58,7 @@
     "@types/react-dom": "^18.0.11",
     "@vitejs/plugin-react": "^3.1.0",
     "@vitejs/plugin-react-swc": "^3.0.0",
+    "contentful": "^10.6.14",
     "husky": "^8.0.3",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
@@ -78,6 +78,7 @@
     "vite-plugin-svgr": "^4.1.0"
   },
   "peerDependencies": {
+    "contentful": ">=10.6.0",
     "react": ">=16.0.0",
     "react-dom": ">=16.0.0"
   },
