+++ packages/experience-builder-sdk/src/ExperienceRoot.tsx
@@ -6,7 +6,7 @@ import {
   isDeprecatedExperience,
   supportedModes,
 } from '@contentful/experience-builder-core';
-import { EntityStore } from './core/preview/EntityStore';
+import { EntityStore } from '@contentful/experience-builder-core';
 import type {
   DeprecatedExperience,
   Experience,
