+++ packages/experience-builder-sdk/src/blocks/preview/CompositionBlock.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import type { UnresolvedLink } from 'contentful';
 import omit from 'lodash.omit';
-import { EntityStore } from '../../core/preview/EntityStore';
+import { EntityStore } from '@contentful/experience-builder-core';
 import {
   CF_STYLE_ATTRIBUTES,
   CONTENTFUL_CONTAINER_ID,
