+++ packages/experience-builder-sdk/src/blocks/preview/DeprecatedPreviewDeliveryRoot.tsx
@@ -41,9 +41,6 @@ export const DeprecatedPreviewDeliveryRoot = ({
   const { fetchBySlug, experience, isFetching } = useFetchExperience({
     client,
     mode: mode as ExternalSDKMode,
-    slug,
-    experienceTypeId,
-    localeCode: locale,
   });
 
   const entityStore = experience?.entityStore;
