+++ packages/experience-builder-sdk/src/blocks/preview/PreviewDeliveryRoot.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { EntityStore } from '../../core/preview/EntityStore';
+import { EntityStore } from '@contentful/experience-builder-core';
 import type { Experience, InternalSDKMode } from '@contentful/experience-builder-core/types';
 import { CompositionBlock } from './CompositionBlock';
 import { compatibleVersions } from '../../constants';
