+++ packages/experience-builder-sdk/src/core/index.ts
@@ -1,2 +1,1 @@
 export { Flex } from '../components/Flex';
-export * as fetchers from './fetchers';
