+++ packages/experience-builder-sdk/src/core/preview/designComponentUtils.spec.ts
@@ -6,7 +6,7 @@ import {
 import { assets, entries } from '../../../test/__fixtures__/entities';
 import { CONTENTFUL_CONTAINER_ID } from '@contentful/experience-builder-core/constants';
 import type { CompositionNode } from '@contentful/experience-builder-core/types';
-import { EntityStore } from './EntityStore';
+import { EntityStore } from '@contentful/experience-builder-core';
 import { resolveDesignComponent } from './designComponentUtils';
 
 describe('resolveDesignComponent', () => {
