+++ packages/visual-editor/package.json
@@ -59,7 +59,7 @@
     "typescript": "^5.0.2"
   },
   "peerDependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": ">=17.0.0",
+    "react-dom": ">=17.0.0"
   }
 }
